Revert "Feature/invite mail"
