+++ src/client/controller/repo.js
@@ -105,14 +105,6 @@ module.controller('RepoCtrl', ['$scope', '$stateParams', '$modal', '$timeout', '
             });
         };
 
-        $scope.inviteEmail = function() {
-            $RPC.call('invitation', 'inviteEmail', {
-                user: $stateParams.user,
-                repo: $stateParams.repo,
-                email: $scope.invitemail
-            });
-        };
-
         //
         // UI text
         //
